reformat factor implement process
