+++ rdagent/factor_implementation/evolving/evolving_strategy.py
@@ -1,59 +1,70 @@
 from __future__ import annotations
 
 import json
-from pathlib import Path
-import random
 from abc import abstractmethod
 from copy import deepcopy
 from typing import TYPE_CHECKING
 
-from rdagent.core.evolving_framework import EvolvingStrategy, QueriedKnowledge
-from rdagent.core.utils import multiprocessing_wrapper
-from rdagent.factor_implementation.share_modules.factor import (
-    FactorImplementation,
-    FactorImplementationTask,
-    FileBasedFactorImplementation,
-)
-from rdagent.core.prompts import Prompts
 from jinja2 import Template
-from rdagent.oai.llm_utils import APIBackend
 
+from rdagent.core.evolving_framework import EvolvingStrategy, QueriedKnowledge
+from rdagent.oai.llm_utils import APIBackend
 from rdagent.factor_implementation.share_modules.factor_implementation_config import (
     FactorImplementSettings,
 )
-from rdagent.factor_implementation.share_modules.factor_implementation_utils import (
-    get_data_folder_intro,
+from rdagent.core.task import (
+    TaskImplementation,
+    FactorImplementTask,
+)
+from rdagent.core.prompts import Prompts
+from pathlib import Path
+from rdagent.factor_implementation.evolving.knowledge_management import FileBasedFactorImplementation
+
+from rdagent.factor_implementation.evolving.scheduler import (
+    RandomSelect,
+    LLMSelect,
 )
 
+from scripts.factor_implementation.share_modules.prompt import (
+    FactorImplementationPrompts,
+)
+
+from rdagent.factor_implementation.share_modules.factor_implementation_utils import get_data_folder_intro
+
+from utils.misc import multiprocessing_wrapper
+
 if TYPE_CHECKING:
-    from factor_implementation.evolving.evolvable_subjects import (
-        FactorImplementationList,
+    from rdagent.core.task import (
+        FactorEvovlingItem,
     )
-    from factor_implementation.evolving.knowledge_management import (
+    from scripts.factor_implementation.baselines.evolving.knowledge_management import (
         FactorImplementationQueriedKnowledge,
         FactorImplementationQueriedKnowledgeV1,
     )
 
+implement_prompts = Prompts(file_path=Path(__file__).parent.parent / "prompts.yaml")
 
 class MultiProcessEvolvingStrategy(EvolvingStrategy):
     @abstractmethod
     def implement_one_factor(
         self,
-        target_task: FactorImplementationTask,
+        target_task: FactorImplementTask,
         queried_knowledge: QueriedKnowledge = None,
-    ) -> FactorImplementation:
+    ) -> TaskImplementation:
         raise NotImplementedError
 
     def evolve(
         self,
         *,
-        evo: FactorImplementationList,
+        evo: FactorEvovlingItem,
         queried_knowledge: FactorImplementationQueriedKnowledge | None = None,
         **kwargs,
-    ) -> FactorImplementationList:
+    ) -> FactorEvovlingItem:
+        self.num_loop += 1
         new_evo = deepcopy(evo)
         new_evo.corresponding_implementations = [None for _ in new_evo.target_factor_tasks]
 
+        # 1.找出需要evolve的factor
         to_be_finished_task_index = []
         for index, target_factor_task in enumerate(new_evo.target_factor_tasks):
             target_factor_task_desc = target_factor_task.get_factor_information()
@@ -66,12 +77,25 @@ class MultiProcessEvolvingStrategy(EvolvingStrategy):
                 and target_factor_task_desc not in queried_knowledge.failed_task_info_set
             ):
                 to_be_finished_task_index.append(index)
-        if FactorImplementSettings().implementation_factors_per_round < len(to_be_finished_task_index):
-            to_be_finished_task_index = random.sample(
-                to_be_finished_task_index,
-                FactorImplementSettings().implementation_factors_per_round,
-            )
 
+        # 2. 选择selection方法
+        # if the number of factors to be implemented is larger than the limit, we need to select some of them
+        if FactorImplementSettings().implementation_factors_per_round < len(to_be_finished_task_index):
+            # if the number of loops is equal to the select_loop, we need to select some of them
+            if FactorImplementSettings().select_method == "random":
+                to_be_finished_task_index = RandomSelect(
+                    to_be_finished_task_index,
+                    FactorImplementSettings().implementation_factors_per_round
+                )
+
+            if FactorImplementSettings().select_method == "scheduler":
+                to_be_finished_task_index = LLMSelect(
+                    to_be_finished_task_index,
+                    FactorImplementSettings().implementation_factors_per_round,
+                    new_evo,
+                    queried_knowledge.all_former_traces,
+                )
+        
         result = multiprocessing_wrapper(
             [
                 (self.implement_one_factor, (new_evo.target_factor_tasks[target_index], queried_knowledge))
@@ -82,11 +106,14 @@ class MultiProcessEvolvingStrategy(EvolvingStrategy):
 
         for index, target_index in enumerate(to_be_finished_task_index):
             new_evo.corresponding_implementations[target_index] = result[index]
+            if result[index].target_task.factor_name in new_evo.evolve_trace:
+                new_evo.evolve_trace[result[index].target_task.factor_name].append(
+                    result[index]
+                )
+            else:
+                new_evo.evolve_trace[result[index].target_task.factor_name] = [result[index]]
 
-        # for target_index in to_be_finished_task_index:
-        #     new_evo.corresponding_implementations[target_index] = self.implement_one_factor(
-        #         new_evo.target_factor_tasks[target_index], queried_knowledge
-        #     )
+        new_evo.corresponding_selection.append(to_be_finished_task_index)
 
         return new_evo
 
@@ -94,9 +121,9 @@ class MultiProcessEvolvingStrategy(EvolvingStrategy):
 class FactorEvolvingStrategy(MultiProcessEvolvingStrategy):
     def implement_one_factor(
         self,
-        target_task: FactorImplementationTask,
+        target_task: FactorImplementTask,
         queried_knowledge: FactorImplementationQueriedKnowledgeV1 = None,
-    ) -> FactorImplementation:
+    ) -> TaskImplementation:
         factor_information_str = target_task.get_factor_information()
 
         if queried_knowledge is not None and factor_information_str in queried_knowledge.success_task_to_knowledge_dict:
@@ -118,9 +145,7 @@ class FactorEvolvingStrategy(MultiProcessEvolvingStrategy):
             queried_former_failed_knowledge_to_render = queried_former_failed_knowledge
 
             system_prompt = Template(
-                Prompts(file_path=Path(__file__).parent.parent / "prompts.yaml")[
-                    "evolving_strategy_factor_implementation_v1_system"
-                ],
+                implement_prompts["evolving_strategy_factor_implementation_v1_system"],
             ).render(
                 data_info=get_data_folder_intro(),
                 queried_former_failed_knowledge=queried_former_failed_knowledge_to_render,
@@ -133,9 +158,7 @@ class FactorEvolvingStrategy(MultiProcessEvolvingStrategy):
             while True:
                 user_prompt = (
                     Template(
-                        Prompts(file_path=Path(__file__).parent.parent / "prompts.yaml")[
-                            "evolving_strategy_factor_implementation_v1_user"
-                        ],
+                        implement_prompts["evolving_strategy_factor_implementation_v1_user"],
                     )
                     .render(
                         factor_information_str=factor_information_str,
@@ -154,9 +177,6 @@ class FactorEvolvingStrategy(MultiProcessEvolvingStrategy):
                     queried_former_failed_knowledge_to_render = queried_former_failed_knowledge_to_render[1:]
                 elif len(queried_similar_successful_knowledge_to_render) > 1:
                     queried_similar_successful_knowledge_to_render = queried_similar_successful_knowledge_to_render[1:]
-            # print(
-            #     f"length of queried_similar_successful_knowledge_to_render: {len(queried_similar_successful_knowledge_to_render)}, length of queried_former_failed_knowledge_to_render: {len(queried_former_failed_knowledge_to_render)}"
-            # )
 
             code = json.loads(
                 session.build_chat_completion(
@@ -172,16 +192,21 @@ class FactorEvolvingStrategy(MultiProcessEvolvingStrategy):
 
             return factor_implementation
 
-
 class FactorEvolvingStrategyWithGraph(MultiProcessEvolvingStrategy):
+    def __init__(self) -> None:
+        self.num_loop = 0
+        self.haveSelected = False
+
     def implement_one_factor(
         self,
-        target_task: FactorImplementationTask,
+        target_task: FactorImplementTask,
         queried_knowledge,
-    ) -> FactorImplementation:
+    ) -> TaskImplementation:
         error_summary = FactorImplementSettings().v2_error_summary
+        # 1. 提取因子的背景信息
         target_factor_task_information = target_task.get_factor_information()
 
+        # 2. 检查该因子是否需要继续做（是否已经作对，是否做错太多）
         if (
             queried_knowledge is not None
             and target_factor_task_information in queried_knowledge.success_task_to_knowledge_dict
@@ -190,6 +215,8 @@ class FactorEvolvingStrategyWithGraph(MultiProcessEvolvingStrategy):
         elif queried_knowledge is not None and target_factor_task_information in queried_knowledge.failed_task_info_set:
             return None
         else:
+
+            # 3. 取出knowledge里面的经验数据（similar success、similar error、former_trace）
             queried_similar_component_knowledge = (
                 queried_knowledge.component_with_success_task[target_factor_task_information]
                 if queried_knowledge is not None
@@ -209,9 +236,7 @@ class FactorEvolvingStrategyWithGraph(MultiProcessEvolvingStrategy):
             queried_former_failed_knowledge_to_render = queried_former_failed_knowledge
 
             system_prompt = Template(
-                Prompts(file_path=Path(__file__).parent.parent / "prompts.yaml")[
-                    "evolving_strategy_factor_implementation_v1_system"
-                ],
+                implement_prompts["evolving_strategy_factor_implementation_v1_system"],
             ).render(
                 data_info=get_data_folder_intro(),
                 queried_former_failed_knowledge=queried_former_failed_knowledge_to_render,
@@ -224,18 +249,17 @@ class FactorEvolvingStrategyWithGraph(MultiProcessEvolvingStrategy):
             queried_similar_component_knowledge_to_render = queried_similar_component_knowledge
             queried_similar_error_knowledge_to_render = queried_similar_error_knowledge
             error_summary_critics = ""
+            # 动态地防止prompt超长
             while True:
+                # 总结error（可选）
                 if (
                     error_summary
                     and len(queried_similar_error_knowledge_to_render) != 0
                     and len(queried_former_failed_knowledge_to_render) != 0
                 ):
+                    
                     error_summary_system_prompt = (
-                        Template(
-                            Prompts(file_path=Path(__file__).parent.parent / "prompts.yaml")[
-                                "evolving_strategy_error_summary_v2_system"
-                            ]
-                        )
+                        Template(implement_prompts["evolving_strategy_error_summary_v2_system"])
                         .render(
                             factor_information_str=target_factor_task_information,
                             code_and_feedback=queried_former_failed_knowledge_to_render[
@@ -249,11 +273,7 @@ class FactorEvolvingStrategyWithGraph(MultiProcessEvolvingStrategy):
                     )
                     while True:
                         error_summary_user_prompt = (
-                            Template(
-                                Prompts(file_path=Path(__file__).parent.parent / "prompts.yaml")[
-                                    "evolving_strategy_error_summary_v2_user"
-                                ]
-                            )
+                            Template(implement_prompts["evolving_strategy_error_summary_v2_user"])
                             .render(
                                 queried_similar_component_knowledge=queried_similar_component_knowledge_to_render,
                             )
@@ -270,12 +290,10 @@ class FactorEvolvingStrategyWithGraph(MultiProcessEvolvingStrategy):
                         user_prompt=error_summary_user_prompt,
                         json_mode=False,
                     )
-
+                # 构建user_prompt。开始写代码
                 user_prompt = (
                     Template(
-                        Prompts(file_path=Path(__file__).parent.parent / "prompts.yaml")[
-                            "evolving_strategy_factor_implementation_v2_user"
-                        ],
+                        implement_prompts["evolving_strategy_factor_implementation_v2_user"],
                     )
                     .render(
                         factor_information_str=target_factor_task_information,
@@ -301,12 +319,6 @@ class FactorEvolvingStrategyWithGraph(MultiProcessEvolvingStrategy):
                     queried_similar_component_knowledge_to_render = queried_similar_component_knowledge_to_render[:-1]
                 elif len(queried_similar_error_knowledge_to_render) > 0:
                     queried_similar_error_knowledge_to_render = queried_similar_error_knowledge_to_render[:-1]
-
-            # print(
-            #     len(queried_similar_component_knowledge_to_render),
-            #     len(queried_similar_error_knowledge_to_render),
-            #     len(queried_former_failed_knowledge_to_render),
-            # )
 
             response = session.build_chat_completion(
                 user_prompt=user_prompt,
