+++ rdagent/factor_implementation/share_modules/factor_implementation_config.py
@@ -1,9 +1,12 @@
 from pathlib import Path
 
-from core.conf import FincoSettings
+from pydantic_settings import BaseSettings
 
+from typing import Literal
 
-class FactorImplementSettings(FincoSettings):
+SELECT_METHOD = Literal["random", "scheduler"]
+
+class FactorImplementSettings(BaseSettings):
     file_based_execution_data_folder: str = str(
         (Path().cwd() / "git_ignore_folder" / "factor_implementation_source_data").absolute(),
     )
@@ -29,14 +32,15 @@ class FactorImplementSettings(FincoSettings):
     v2_error_summary: bool = False
     v2_knowledge_sampler: float = 1.0
 
-    chat_token_limit: int = (
-        100000  # 100000 is the maximum limit of gpt4, which might increase in the future version of gpt
-    )
-
     implementation_factors_per_round: int = 100  # how many factors to choose for each round of evolving
     evo_multi_proc_n: int = 16  # how many processes to use for evolving (including eval & generation)
 
     file_based_execution_timeout: int = 120  # seconds for each factor implementation execution
 
+    select_method: SELECT_METHOD = "random"
+
+    chat_token_limit: int = (
+        100000  # 100000 is the maximum limit of gpt4, which might increase in the future version of gpt
+    )
 
 FIS = FactorImplementSettings()
