+++ rdagent/factor_implementation/share_modules/factor_implementation_utils.py
@@ -6,6 +6,8 @@ import pandas as pd
 from jinja2 import Template
 
 from rdagent.factor_implementation.share_modules.factor_implementation_config import FIS
+from rdagent.core.task import FactorImplementTask
+
 
 TPL = """
 {{file_name}}
@@ -15,7 +17,6 @@ TPL = """
 """
 # Create a Jinja template from the string
 JJ_TPL = Template(TPL)
-
 
 def get_data_folder_intro():
     """Direclty get the info of the data folder.
@@ -49,3 +50,17 @@ def get_data_folder_intro():
                 f"file type {p.name} is not supported. Please implement its description function.",
             )
     return "\n ----------------- file spliter -------------\n".join(content_l)
+
+def load_data_from_dict(factor_dict:dict) -> list[FactorImplementTask]:
+    """Load data from a dict.
+    """
+    task_l = []
+    for factor_name, factor_data in factor_dict.items():
+        task = FactorImplementTask(
+            factor_name=factor_name,
+            factor_description=factor_data["description"],
+            factor_formulation=factor_data["formulation"],
+            variables=factor_data["variables"],
+        )
+        task_l.append(task)
+    return task_l
