+++ rdagent/oai/llm_utils.py
@@ -16,7 +16,7 @@ from typing import List, Optional, Tuple, Union
 import numpy as np
 import tiktoken
 
-from rdagent.core.conf import FincoSettings as Config
+from rdagent.core.conf import RDAgentSettings as Config
 from rdagent.core.log import FinCoLog, LogColors
 from rdagent.core.utils import SingletonBaseClass
